requirements.txt
absl-py==2.3.0
addict==2.4.0
aiofiles==22.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.8
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
beautifulsoup4==4.13.4
black==23.11.0
bleach==6.2.0
braceexpand==0.1.7
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
coverage==7.9.1
cycler==0.12.1
datasets==2.21.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
executing==2.2.0
fastapi==0.115.14
fastjsonschema==2.21.1
ffmpeg-python==0.2.0
ffmpy==0.6.0
filelock==3.18.0
flake8==6.1.0
fonttools==4.58.4
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
gradio==5.35.0
gradio_client==1.10.4
groovy==0.1.2
grpcio==1.73.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imgaug==0.4.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
isort==6.0.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
lazy_loader==0.4
librosa==0.11.0
lightning==2.5.2
lightning-utilities==0.14.3
llvmlite==0.42.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
memory-profiler==0.61.0
mistune==3.1.3
mmcv==2.1.0
mmdet==3.3.0
mmengine==0.10.7
mpmath==1.3.0
msgpack==1.1.1
multidict==6.5.0
multiprocess==0.70.16
mypy==1.16.1
mypy_extensions==1.1.0
narwhals==1.43.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.4.3
notebook_shim==0.2.4
numba==0.59.1
numpy==2.2.6
omegaconf==2.3.0
opencv-python==4.12.0.88
opencv-python-headless==4.11.0.86
orjson==3.10.18
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.8
plotly==6.1.2
pluggy==1.6.0
pooch==1.8.2
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pycocotools==2.0.10
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pyflakes==3.1.0
Pygments==2.19.2
pyparsing==3.2.3
pytest==7.4.3
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytorch-lightning==2.5.2
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
qudida==0.0.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.25.1
ruff==0.12.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentry-sdk==2.30.0
setproctitle==1.3.6
shapely==2.1.1
shellingham==1.5.4
simsimd==6.4.9
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
stack-data==0.6.3
starlette==0.46.2
stringzilla==3.12.5
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminado==0.18.1
terminaltables==3.1.10
threadpoolctl==3.6.0
tifffile==2025.6.11
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.1
tomlkit==0.12.0
torch==2.7.1
torchaudio==2.7.1
torchmetrics==1.7.3
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
typer==0.16.0
types-python-dateutil==2.9.0.20250516
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.24.0
wandb==0.20.1
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webdataset==1.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
widgetsnbextension==4.0.14
xxhash==3.5.0
yapf==0.43.0
yarl==1.20.1
yt-dlp==2025.6.30